requirements.txt
# 웹 프레임워크
fastapi>=0.104.0
uvicorn>=0.24.0

# 데이터베이스
sqlalchemy
psycopg2-binary
jaydebeapi>=1.2.3

# 데이터 검증 및 설정
pydantic
pydantic-settings
email-validator>=2.0.0

# 비밀번호 해싱
bcrypt
passlib[bcrypt]

# AI 및 머신러닝
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
instructor>=0.4.0

# 이미지 처리
opencv-python>=4.8.0
pillow>=10.0.0

# 수치 계산 및 시각화
numpy>=1.24.0
matplotlib>=3.7.0

# 웹 크롤링
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# 기타 유틸리티
python-multipart
python-dotenv>=1.0.0
google-generativeai>=0.3.0